Move GetKubeDirectorReference() to observer.go.
